+++ Worker/WorkflowTaskCoordinator.cs
@@ -1,15 +1,9 @@
 ﻿using Conductor.Client.Interfaces;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Net.Http.Json;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace Conductor.Client.Worker
@@ -21,26 +15,30 @@ namespace Conductor.Client.Worker
         private ILogger<WorkflowTaskCoordinator> logger;
         private HashSet<Type> workerDefinitions = new HashSet<Type>();
         private ConductorWorkerClientConfiguration conductorClientSetting;
+        private ConductorAuthTokenClient conductorAuthTokenClient;
 
         public WorkflowTaskCoordinator(IServiceProvider serviceProvider,
             ILogger<WorkflowTaskCoordinator> logger,
-            IOptions<ConductorWorkerClientConfiguration> conductorClientSettings)
+            IOptions<ConductorWorkerClientConfiguration> conductorClientSettings,
+            ConductorAuthTokenClient conductorAuthTokenClient)
          {
             this.serviceProvider = serviceProvider;
             this.logger = logger;
             concurrentWorkers = conductorClientSettings.Value.ConcurrentWorkers;
             conductorClientSetting = conductorClientSettings.Value;
+            this.conductorAuthTokenClient = conductorAuthTokenClient;
+
         }
 
         public async Task Start()
         {
             logger.LogInformation("Starting WorkflowCoordinator");
-            if (this.conductorClientSetting.AuthenticationClient != null 
-                && !string.IsNullOrEmpty(this.conductorClientSetting.AuthenticationClient.keyId)
-                && !string.IsNullOrEmpty(this.conductorClientSetting.AuthenticationClient.keySecret))
+            if (this.conductorClientSetting.AuthenticationConfiguration != null 
+                && !string.IsNullOrEmpty(this.conductorClientSetting.AuthenticationConfiguration.keyId)
+                && !string.IsNullOrEmpty(this.conductorClientSetting.AuthenticationConfiguration.keySecret))
             {
-                this.conductorClientSetting.Token = PostForToken(this.conductorClientSetting.AuthenticationClient.keyId,
-                                                                 this.conductorClientSetting.AuthenticationClient.keySecret).Result;
+                this.conductorClientSetting.Token = this.conductorAuthTokenClient.PostForToken(this.conductorClientSetting.ServerUrl + "/token", this.conductorClientSetting.AuthenticationConfiguration.keyId,
+                                                                 this.conductorClientSetting.AuthenticationConfiguration.keySecret).Result;
             }
             var pollers = new List<Task>();
             for (var i = 0; i < concurrentWorkers; i++)
@@ -56,27 +54,6 @@ namespace Conductor.Client.Worker
         {
             workerDefinitions.Add(task.GetType());
         }
-        public async Task<string> PostForToken(String keyId, String keySecret)
-        {
-            HttpClient httpClient = new HttpClient();
-            var urlBuilder = new StringBuilder( this.conductorClientSetting.ServerUrl + "/token");
 
-            using (var request = new HttpRequestMessage { Method = System.Net.Http.HttpMethod.Post, RequestUri = new Uri(urlBuilder.ToString(), UriKind.RelativeOrAbsolute) })
-            {
-                request.Headers.Accept.Add(MediaTypeWithQualityHeaderValue.Parse("application/json"));
-                request.Content = JsonContent.Create(new
-                {
-                    keyId = keyId,
-                    keySecret = keySecret,
-                });
-                var response = httpClient.SendAsync(request, HttpCompletionOption.ResponseHeadersRead).Result;
-
-
-                var result = (JObject)JsonConvert.DeserializeObject(response.Content.ReadAsStringAsync().Result);
-
-                return result["token"].Value<string>();
-
-            }
-        }
     }
 }
