+++ conductor-csharp.csproj
@@ -4,7 +4,7 @@
     <TargetFramework>netstandard2.0</TargetFramework>
     <Version>0.0.2</Version>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageIcon>Orkes.jpg</PackageIcon>
+    <PackageIcon>avatar.jpg</PackageIcon>
     <PackageIconUrl />
   </PropertyGroup>
 
@@ -12,6 +12,10 @@
     <Compile Remove="conductor-csharp.Test\**" />
     <EmbeddedResource Remove="conductor-csharp.Test\**" />
     <None Remove="conductor-csharp.Test\**" />
+    <None Include="..\..\avatar.jpg">
+      <Pack>True</Pack>
+      <PackagePath></PackagePath>
+    </None>
     <None Include="..\..\Users\manan\Downloads\Orkes.jpg">
       <Pack>True</Pack>
       <PackagePath></PackagePath>
