+++ conductor-csharp.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 16.0.31313.79
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "conductor-csharp", "conductor-csharp.csproj", "{EB44C772-CDAB-4C91-9B8F-BF0241DF1CC6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "conductor-csharp.tests", "..\conductor-csharp.tests\conductor-csharp.tests.csproj", "{64FBF720-45EC-4AF5-B3ED-D95414AFA44C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "conductor-csharp.tests", "..\conductor-csharp.tests\conductor-csharp.tests.csproj", "{64FBF720-45EC-4AF5-B3ED-D95414AFA44C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleApp3", "..\ConsoleApp3\ConsoleApp3.csproj", "{C9EDF2D5-AA9D-47CE-AE92-8A37D4BFCB40}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{64FBF720-45EC-4AF5-B3ED-D95414AFA44C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{64FBF720-45EC-4AF5-B3ED-D95414AFA44C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{64FBF720-45EC-4AF5-B3ED-D95414AFA44C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C9EDF2D5-AA9D-47CE-AE92-8A37D4BFCB40}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C9EDF2D5-AA9D-47CE-AE92-8A37D4BFCB40}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C9EDF2D5-AA9D-47CE-AE92-8A37D4BFCB40}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C9EDF2D5-AA9D-47CE-AE92-8A37D4BFCB40}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
